+++ webknossos/webknossos/dataset/_array.py
@@ -430,7 +430,7 @@ class TensorStoreArray(BaseArray):
             parsed_url = urlparse(str(path))
             kvstore_spec: dict[str, Any] = {
                 "driver": "s3",
-                "path": parsed_url.path,
+                "path": parsed_url.path.lstrip("/"),
                 "bucket": parsed_url.netloc,
             }
             if endpoint_url := path.storage_options.get("client_kwargs", {}).get(