+++ webknossos/Changelog.md
@@ -154,6 +154,7 @@ For upgrade instructions, please check the respective _Breaking Changes_ section
 
 
 ### Added
+- Added the `webknossos copy-dataset` CLI command. [#1259](https://github.com/scalableminds/webknossos-libs/pull/1259)
 - Added `Dataset.write_layer` method for writing entire layers in one go. [#1242](https://github.com/scalableminds/webknossos-libs/pull/1242)
 
 ### Changed